requirements.txt
Flask
pandas
PyPDF2
python-docx
